expense

+++ Account/database/migrations/2021_04_18_180658_create_expenses_table.php
@@ -19,7 +19,7 @@ class CreateExpensesTable extends Migration
             $table->string('number');
             $table->date('date_issue');
             $table->string('seller');
-            $table->integer('nip');
+            $table->bigInteger('nip');
             $table->string('name');
             $table->double('netto');
             $table->integer('vat');
